requirements.txt
flask
pymongo
python-dotenv
dnspython
pandas
sklearn
nltk